api/requirements.txt
Django>=2.0,<2.1
djangorestframework>=3.7,<3.8
drf-yasg
